chore(sdk): Split `linked_chunk` into 2 modules.
